infra/aws/cdk/requirements.txt
aws-cdk-lib>=2.140
constructs>=10
